mark tests that use external resources as integration tests
