+++ pkg/install/images/integration/testdata/valid-archive.tar.gz
